Handle retries, delay queue and quarantine
