+++ src/main/java/uk/gov/ons/census/notifyprocessor/config/MessageConsumerConfig.java
@@ -1,0 +1,129 @@
+package uk.gov.ons.census.notifyprocessor.config;
+
+import org.springframework.amqp.rabbit.config.RetryInterceptorBuilder;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.amqp.rabbit.listener.SimpleMessageListenerContainer;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.integration.amqp.inbound.AmqpInboundChannelAdapter;
+import org.springframework.integration.channel.DirectChannel;
+import org.springframework.messaging.MessageChannel;
+import org.springframework.retry.backoff.FixedBackOffPolicy;
+import org.springframework.retry.interceptor.RetryOperationsInterceptor;
+import uk.gov.ons.census.notifyprocessor.client.ExceptionManagerClient;
+import uk.gov.ons.census.notifyprocessor.messaging.ManagedMessageRecoverer;
+import uk.gov.ons.census.notifyprocessor.model.EnrichedFulfilmentRequest;
+import uk.gov.ons.census.notifyprocessor.model.ResponseManagementEvent;
+
+@Configuration
+public class MessageConsumerConfig {
+  private final ExceptionManagerClient exceptionManagerClient;
+  private final RabbitTemplate rabbitTemplate;
+  private final ConnectionFactory connectionFactory;
+
+  @Value("${messagelogging.logstacktraces}")
+  private boolean logStackTraces;
+
+  @Value("${queueconfig.consumers}")
+  private int consumers;
+
+  @Value("${queueconfig.retry-attempts}")
+  private int retryAttempts;
+
+  @Value("${queueconfig.retry-delay}")
+  private int retryDelay;
+
+  @Value("${queueconfig.retry-exchange}")
+  private String retryExchange;
+
+  @Value("${queueconfig.quarantine-exchange}")
+  private String quarantineExchange;
+
+  @Value("${queueconfig.fulfilment-request-inbound-queue}")
+  private String fulfilmentInboundQueue;
+
+  @Value("${queueconfig.enriched-fulfilment-queue}")
+  private String enrichedFulfilmentQueue;
+
+  public MessageConsumerConfig(
+      ExceptionManagerClient exceptionManagerClient,
+      RabbitTemplate rabbitTemplate,
+      ConnectionFactory connectionFactory) {
+    this.exceptionManagerClient = exceptionManagerClient;
+    this.rabbitTemplate = rabbitTemplate;
+    this.connectionFactory = connectionFactory;
+  }
+
+  @Bean
+  public MessageChannel fulfilmentInputChannel() {
+    return new DirectChannel();
+  }
+
+  @Bean
+  public MessageChannel enrichedFulfilmentInputChannel() {
+    return new DirectChannel();
+  }
+
+  @Bean
+  public AmqpInboundChannelAdapter fulfilmentInbound(
+      @Qualifier("fulfilmentContainer") SimpleMessageListenerContainer listenerContainer,
+      @Qualifier("fulfilmentInputChannel") MessageChannel channel) {
+    AmqpInboundChannelAdapter adapter = new AmqpInboundChannelAdapter(listenerContainer);
+    adapter.setOutputChannel(channel);
+    return adapter;
+  }
+
+  @Bean
+  public AmqpInboundChannelAdapter enrichedfulfilmentInbound(
+      @Qualifier("enrichedFulfilmentContainer") SimpleMessageListenerContainer listenerContainer,
+      @Qualifier("enrichedFulfilmentInputChannel") MessageChannel channel) {
+    AmqpInboundChannelAdapter adapter = new AmqpInboundChannelAdapter(listenerContainer);
+    adapter.setOutputChannel(channel);
+    return adapter;
+  }
+
+  @Bean
+  public SimpleMessageListenerContainer fulfilmentContainer() {
+    return setupListenerContainer(fulfilmentInboundQueue, ResponseManagementEvent.class);
+  }
+
+  @Bean
+  public SimpleMessageListenerContainer enrichedFulfilmentContainer() {
+    return setupListenerContainer(enrichedFulfilmentQueue, EnrichedFulfilmentRequest.class);
+  }
+
+  private SimpleMessageListenerContainer setupListenerContainer(
+      String queueName, Class expectedMessageType) {
+    FixedBackOffPolicy fixedBackOffPolicy = new FixedBackOffPolicy();
+    fixedBackOffPolicy.setBackOffPeriod(retryDelay);
+
+    ManagedMessageRecoverer managedMessageRecoverer =
+        new ManagedMessageRecoverer(
+            exceptionManagerClient,
+            expectedMessageType,
+            logStackTraces,
+            "Notify Processor",
+            queueName,
+            retryExchange,
+            quarantineExchange,
+            rabbitTemplate);
+
+    RetryOperationsInterceptor retryOperationsInterceptor =
+        RetryInterceptorBuilder.stateless()
+            .maxAttempts(retryAttempts)
+            .backOffPolicy(fixedBackOffPolicy)
+            .recoverer(managedMessageRecoverer)
+            .build();
+
+    SimpleMessageListenerContainer container =
+        new SimpleMessageListenerContainer(connectionFactory);
+    container.setQueueNames(queueName);
+    container.setConcurrentConsumers(consumers);
+    container.setChannelTransacted(true);
+    container.setAdviceChain(retryOperationsInterceptor);
+    return container;
+  }
+}
