+++ src/main/java/uk/gov/ons/census/notifyprocessor/model/ExceptionReportResponse.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ExceptionReportResponse {
-  private boolean peek = false;
-  private boolean logIt = false;
-  private boolean skipIt = false;
+  private boolean peek;
+  private boolean logIt;
+  private boolean skipIt;
 }
