+++ src/main/resources/application.yml
@@ -13,6 +13,10 @@ queueconfig:
   enriched-fulfilment-exchange: notify.enriched.fulfilment.exchange
   enriched-fulfilment-queue: notify.enriched.fulfilment
   consumers: 50
+  retry-attempts: 3
+  retry-delay: 1000 #milliseconds
+  retry-exchange: delayedRedeliveryExchange
+  quarantine-exchange: quarantineExchange
 
 healthcheck:
   frequency: 1000 #milliseconds
